pyproject.toml
[project]
name = "mycrew"
version = "0.1.0"
description = "mycrew using crewAI"
authors = [{ name = "Fabrice Fourel", email = "ffourel@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "click>=8.3.0",
    "crewai[google-genai]>=1.4.1",
    "crewai-tools>=1.4.1",
    "fastapi[standard]>=0.121.1",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "google-generativeai>=0.8.5",
    "markdown>=3.10",
    "pathlib2>=2.3.7.post1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "urllib3>=2.3.0",
    "slugify>=0.0.1",
]



[tool.crewai]
type = "crew"
